Refactor app handler
