--- backend/interfaces/api/server/httputils/error.go
@@ -1,10 +1,0 @@
-package httputils
-
-type Error struct {
-	Status int
-	Err    error
-}
-
-func (e Error) Error() string {
-	return e.Err.Error()
-}
