+++ backend/interfaces/api/server/middleware/logger.go
@@ -10,7 +10,7 @@ import (
 func LoggerMiddleware(logger logger.Logger) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			logger.Info("request info",
+			logger.Info("request",
 				zap.String("method", r.Method),
 				zap.String("url", r.URL.Path))
 			next.ServeHTTP(w, r)
